Disable BraveSearchTest browser tests for now
These tests from brave_search_browsertest.cc are failing now on
DCHECK-enabled builds because of a recent change in V8 that adds
an extra check, which seems to require adapting the tests.

We filed https://github.com/brave/brave-browser/issues/16006 to
track fixing this, but let's disable them in the meantime as this
seems to be a test-related issue only (not production-related).

V8 change:

https://chromium.googlesource.com/v8/v8/+/8f65f82033c4dcee7816225cf66340e0d19cae73

commit 8f65f82033c4dcee7816225cf66340e0d19cae73
Author: Jochen Eisinger <jochen@chromium.org>
Date:   Thu May 6 11:06:58 2021 +0200

    Remove flag to disable microtasks scope consistency checks

    Bug: chromium:728583