[FOLLOWUP] Adapt to supporting configurable key storage on linux.
The patch of the application name might not be neccessary any more.
Instead the app name can be set via key_storage_config_linux
(application_name), but there is a comment indicating that this would
only work when allow_runtime_configurable_key_storage feature is
enabled, so need to follow up on that.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/670187f30dbd769f7ad3cdabdd5bbff06412f3e7

commit 670187f30dbd769f7ad3cdabdd5bbff06412f3e7
Author: Samuel Attard <samuel.r.attard@gmail.com>
Date:   Wed May 19 23:15:45 2021 +0000

    support runtime configurable key storage on linux (os_crypto)

    This modifies the OsCrypt::Config struct used on linux to support
    runtime configurable application names which are used in the Keyring and
    LibSecret implementations of os_crypt on linux.